Implement rudimentary `create()` function
Currently placeholder values are used for the password and the new
`User` to be inserted into the `HashMap`. The `HashMap` read and write
from JSON was successful through `serde`, but there is an issue with the
design of the program. We should be inserting a `(String, String)` tuple
hash into the `HashMap`.