Fully refactor all functions to work with the new error handling system excepting for unimplemented functions
Refactoring all of the functions in this manner unveiled a lot of shoddy
error propagating. Most notably this helped reveal a problem in the
method of insertion into the HashMap which allowed the user profiles to
be overwritten upon insert, and even worse, for a successful insertion
to be treated as an error.

-	Separate tests into a new `mod` to help organize in preparation
	for new tests.
-	Derive `Clone` for `User` and `Stock` to allow for error
	reporting if needed.
-	Revise `ProjectError` variants & their implementations.