Refactor most user-methods to properly notify the user of success of their command
Code refactor to user a new `notify()` method to centralize small
successful operation notifications in case this behavior needs to change
in the future. Design changed for `buy_stock()` to provide a better
error to the user when generating a parsing error by providing a
`String` insteady of a `u32` for the quantity of stocks to be purchased.